[webgpu] Use i32 as coords type in WGSL (#5628)

+++ tfjs-backend-webgpu/src/kernels/pool2d_webgpu.ts
@@ -137,7 +137,7 @@ export class Pool2DProgram implements WebGPUProgram {
                 continue;
               }
 
-              let value = getX(batch, u32(xR), u32(xC), coords[3]);
+              let value = getX(batch, xR, xC, coords[3]);
               ${updateSnippet}
             }
           }
